Rename coroutine to coro: includes
