+++ include/yaclib/coro/async_mutex.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <yaclib/coroutine/coroutine.hpp>
-#include <yaclib/coroutine/detail/promise_type.hpp>
+#include <yaclib/coro/coroutine.hpp>
+#include <yaclib/coro/detail/promise_type.hpp>
 #include <yaclib/executor/inline.hpp>
 #include <yaclib/executor/strand.hpp>
 #include <yaclib/executor/thread_pool.hpp>
