+++ include/yaclib/coro/detail/await_awaiter.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <yaclib/async/future.hpp>
-#include <yaclib/coroutine/coroutine.hpp>
+#include <yaclib/coro/coroutine.hpp>
 #include <yaclib/executor/thread_pool.hpp>
 #include <yaclib/util/detail/atomic_counter.hpp>
 #include <yaclib/util/type_traits.hpp>
