+++ include/yaclib/coro/future_traits.hpp
@@ -2,7 +2,7 @@
 
 #include <yaclib/async/future.hpp>
 #include <yaclib/async/task.hpp>
-#include <yaclib/coroutine/detail/promise_type.hpp>
+#include <yaclib/coro/detail/promise_type.hpp>
 
 /**
  * TODO(mkornaukhov03) Add doxygen docs
