+++ test/unit/coro/await_group_stress.cpp
@@ -2,9 +2,9 @@
 
 #include <yaclib/algo/wait_group.hpp>
 #include <yaclib/async/run.hpp>
-#include <yaclib/coroutine/await.hpp>
-#include <yaclib/coroutine/future_traits.hpp>
-#include <yaclib/coroutine/on.hpp>
+#include <yaclib/coro/await.hpp>
+#include <yaclib/coro/future_traits.hpp>
+#include <yaclib/coro/on.hpp>
 #include <yaclib/executor/manual.hpp>
 #include <yaclib/executor/thread_pool.hpp>
